refactor : templates 추가 및 radio group 추가
