+++ components/Templates/Cardview/index.ts
@@ -1,0 +1,5 @@
+import { AllCard } from './AllCard';
+import { CardCategory } from './CardCategory';
+import { PopularCard } from './PopularCard';
+
+export { AllCard, CardCategory, PopularCard };
