+++ pages/_app.tsx
@@ -4,6 +4,7 @@ import type { AppProps } from 'next/app';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import CommonLayout from '@/components/common/CommonLayout';
+import { Header } from '@/components/Atom/Header';
 import { RecoilRoot } from 'recoil';
 
 function MyApp({ Component, pageProps }: AppProps) {
@@ -12,6 +13,7 @@ function MyApp({ Component, pageProps }: AppProps) {
     <RecoilRoot>
       <QueryClientProvider client={queryClient}>
         <GlobalStyles />
+        <Header isLogin={true}></Header>
         <CommonLayout>
           <Component {...pageProps} />
         </CommonLayout>
